Rename useSearchParams to useQueryParams
