+++ components/movies/Discover.tsx
@@ -2,7 +2,7 @@ import type { MouseEvent } from 'react';
 
 import useSWR from 'swr';
 import { useRouter } from 'next/router';
-import { useQueryParams } from '@/hooks/useSearchParams';
+import { useQueryParams } from '@/hooks/useQueryParams';
 
 import MovieList from '../common/movies/MovieList';
 import MovieListContainer from '../common/movies/MovieListContainer';
